// constants/types.ts
export type Apartments = {
    available_from?: string | null;
    building?: string | Buildings | null;
    date_created?: string | null;
    date_updated?: string | null;
    external_identifier?: string | null;
    family_friendly?: boolean | null;
    handicapped_accessible?: boolean | null;
    id: string;
    singleflat?: boolean | null;
    sort?: number | null;
    status?: string | null;
    user_created?: string | DirectusUsers | null;
    user_updated?: string | DirectusUsers | null;
    washingmachines: any[] | Washingmachines[];
  };
// ...
  export type Buildings = {
    alias?: string | null;
    apartments: any[] | Apartments[];
    businesshours: any[] | BuildingsBusinesshours[];
    coordinates?: unknown | null;
    date_created?: string | null;
    date_of_construction?: string | null;
    date_updated?: string | null;
    external_identifier?: string | null;
    id: string;
    image?: string | DirectusFiles | null;
    image_remote_url?: string | null;
    image_thumb_hash?: string | null;
    sort?: number | null;
    status?: string | null;
    translations: any[] | BuildingsTranslations[];
    url?: string | null;
    user_created?: string | DirectusUsers | null;
    user_updated?: string | DirectusUsers | null;
  };
  
  export type BuildingsBusinesshours = {
    buildings_id?: string | Buildings | null;
    businesshours_id?: string | Businesshours | null;
    id: number;
  };
  
  export type BuildingsTranslations = {
    be_source_for_translations?: boolean | null;
    buildings_id?: string | Buildings | null;
    content?: string | null;
    id: number;
    languages_code?: string | Languages | null;
    let_be_translated?: boolean | null;
    translation_settings: string;
  };
  
  export type Businesshours = {
    date_created?: string | null;
    date_updated?: string | null;
    date_valid_from?: string | null;
    date_valid_till?: string | null;
    friday?: boolean | null;
    id: string;
    monday?: boolean | null;
    saturday?: boolean | null;
    sort?: number | null;
    status?: string | null;
    sunday?: boolean | null;
    thursday?: boolean | null;
    time_end?: string | null;
    time_start?: string | null;
    tuesday?: boolean | null;
    user_created?: string | DirectusUsers | null;
    user_updated?: string | DirectusUsers | null;
    valid_days: string;
    valid_range: string;
    wednesday?: boolean | null;
  };
// ...
  export type Canteens = {
    alias?: string | null;
    building?: string | Buildings | null;
    date_created?: string | null;
    date_updated?: string | null;
    external_identifier?: string | null;
    foodoffers_import_delete_all_without_dates?: boolean | null;
    foodoffers_import_without_date?: boolean | null;
    foodservice_hours: any[] | CanteensFoodserviceHours[];
    foodservice_hours_during_semester_break: any[] | CanteensFoodserviceHoursDuringSemesterBreak[];
    id: string;
    sort?: number | null;
    status?: string | null;
    user_created?: string | DirectusUsers | null;
    user_updated?: string | DirectusUsers | null;
    utilization_group?: string | UtilizationsGroups | null;
  };
// ...
  export type CanteensFoodserviceHours = {
    businesshours_id?: string | Businesshours | null;
    canteens_id?: string | Canteens | null;
    id: number;
  };
  
  export type CanteensFoodserviceHoursDuringSemesterBreak = {
    businesshours_id?: string | Businesshours | null;
    canteens_id?: string | Canteens | null;
    id: number;
  };
// ...
  export type DirectusAccess = {
    id: string;
    policy: string | DirectusPolicies;
    role?: string | DirectusRoles | null;
    sort?: number | null;
    user?: string | DirectusUsers | null;
  };
// ...
  export type DirectusFiles = {
    charset?: string | null;
    created_on: string;
    description?: string | null;
    duration?: number | null;
    embed?: string | null;
    filename_disk?: string | null;
    filename_download: string;
    filesize?: number | null;
    focal_point_x?: number | null;
    focal_point_y?: number | null;
    folder?: string | DirectusFolders | null;
    height?: number | null;
    id: string;
    location?: string | null;
    metadata?: unknown | null;
    modified_by?: string | DirectusUsers | null;
    modified_on: string;
    storage: string;
    tags?: unknown | null;
    title?: string | null;
    tus_data?: unknown | null;
    tus_id?: string | null;
    type?: string | null;
    uploaded_by?: string | DirectusUsers | null;
    uploaded_on?: string | null;
    width?: number | null;
  };
// ...
  export type DirectusFolders = {
    id: string;
    name: string;
    parent?: string | DirectusFolders | null;
  };
// ...
  export type DirectusRoles = {
    children: any[] | DirectusRoles[];
    description?: string | null;
    icon: string;
    id: string;
    name: string;
    parent?: string | DirectusRoles | null;
    policies: any[] | DirectusAccess[];
    users: any[] | DirectusUsers[];
    users_group: string;
  };
// ...
  export type DirectusUsers = {
    appearance?: string | null;
    auth_data?: unknown | null;
    avatar?: string | DirectusFiles | null;
    description?: string | null;
    email?: string | null;
    email_notifications?: boolean | null;
    external_identifier?: string | null;
    first_name?: string | null;
    id: string;
    language?: string | null;
    last_access?: string | null;
    last_name?: string | null;
    last_page?: string | null;
    location?: string | null;
    password?: string | null;
    policies: any[] | DirectusAccess[];
    profile?: string | Profiles | null;
    provider: string;
    role?: string | DirectusRoles | null;
    status: string;
    tags?: unknown | null;
    tfa_secret?: string | null;
    theme_dark?: string | null;
    theme_dark_overrides?: unknown | null;
    theme_light?: string | null;
    theme_light_overrides?: unknown | null;
    title?: string | null;
    token?: string | null;
  };
// ...
  export type Languages = {
    code: string;
    direction?: string | null;
    name?: string | null;
    status?: string | null;
  };
// ...
  export type Profiles = {
    avatar?: unknown | null;
    buildings_favorites: any[] | ProfilesBuildingsFavorites[];
    buildings_last_opened: any[] | ProfilesBuildingsLastOpened[];
    canteen?: string | Canteens | null;
    course_timetable?: string | null;
    credit_balance?: number | null;
    credit_balance_date_updated?: string | null;
    credit_balance_last_transaction?: number | null;
    date_created?: string | null;
    date_privacy_policy_accepted?: string | null;
    date_updated?: string | null;
    devices: any[] | Devices[];
    foods_feedbacks: any[] | FoodsFeedbacks[];
    id: string;
    language?: string | Languages | null;
    markings: any[] | ProfilesMarkings[];
    nickname?: string | null;
    notifiy_on_free_apartments?: boolean | null;
    price_group?: string | null;
    sort?: number | null;
    status?: string | null;
    user_created?: string | DirectusUsers | null;
    user_updated?: string | DirectusUsers | null;
  };
// ...
  export type UtilizationsEntries = {
    date_created?: string | null;
    date_end?: string | null;
    date_start?: string | null;
    date_updated?: string | null;
    id: string;
    sort?: number | null;
    status?: string | null;
    user_created?: string | DirectusUsers | null;
    user_updated?: string | DirectusUsers | null;
    utilization_group?: string | UtilizationsGroups | null;
    value_forecast_current?: number | null;
    value_real?: number | null;
  };
  
  export type UtilizationsGroups = {
    alias?: string | null;
    all_time_high?: number | null;
    date_created?: string | null;
    date_updated?: string | null;
    id: string;
    sort?: number | null;
    status?: string | null;
    threshold_until_high?: number | null;
    threshold_until_low?: number | null;
    threshold_until_max?: number | null;
    threshold_until_medium?: number | null;
    user_created?: string | DirectusUsers | null;
    user_updated?: string | DirectusUsers | null;
    utilization_entries: any[] | UtilizationsEntries[];
  };
  
  export type Washingmachines = {
    alias?: string | null;
    apartment?: string | Apartments | null;
    date_created?: string | null;
    date_finished?: string | null;
    date_stated?: string | null;
    date_updated?: string | null;
    external_identifier?: string | null;
    id: string;
    sort?: number | null;
    status?: string | null;
    user_created?: string | DirectusUsers | null;
    user_updated?: string | DirectusUsers | null;
  };